JournayPage components moved to folder <journey>
