--- components/Header-plug/Header.module.css
@@ -1,6 +1,0 @@
-.header_plug {
-    display: block;
-    width: 100vw;
-    max-height: 68px;
-    margin-bottom: 24px;
-}
