--- components/Header-plug/Header.tsx
@@ -1,10 +1,0 @@
-import css from "./Header.module.css";
-
-export default function Header() {
-  return (
-    <header className={css.header_plug}>
-      <p>Stork App</p>
-      <hr />
-    </header>
-  );
-}
